SuiteDispatcher was duplicated
